backend/requirements.txt
Django==5.1.2
channels==4.1.0
daphne==4.1.2
psycopg-binary==3.2.10
python-dotenv==1.1.1
pgvector==0.4.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
Pillow==12.0.0
pyyaml==6.0.3
